+++ example/package.json
@@ -10,12 +10,16 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~3.2.3",
+    "@react-navigation/native": "^6.1.18",
+    "@react-navigation/native-stack": "^6.11.0",
     "@wayne-kim/react-native-layout": "^0.3.3",
     "expo": "~51.0.28",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.74.5",
+    "react-native-safe-area-context": "4.10.5",
+    "react-native-screens": "3.31.1",
     "react-native-web": "~0.19.10"
   },
   "devDependencies": {
